[#57] refactor: SessionKeys를 web package로 이동
